Final commit before hiatus.
This project, together with the accompanying frontend, existence, is just a rework of an already working project, bilog, albeit with much better architecture and features. However, due to the relation between the two pairs, I come to a conclusion that the learning that I will get from making these reworks are minimal when compared to creating a brand new project. As such, I will be committing these finals codes to the repository and move on to greener fields. If an abundance of free time comes in the future, I may get back to them again.